pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resumex-py"
version = "1.0.0"
description = "YAML -> LaTeX -> PDF resume builder with strict theme schemas."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "dote"}]
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "typer>=0.12",
  "rich>=13.7",
]
keywords = ["resume", "generator", "latex", "jinja2", "yaml", "pydantic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
]

[project.urls]
Homepage = "https://github.com/rsm-vsivasankaran/resumex"
Repository = "https://github.com/rsm-vsivasankaran/resumex"
Issues = "https://github.com/rsm-vsivasankaran/resumex/issues"

[project.scripts]
resumex = "resumex.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/resumex"]

[tool.hatch.build]
include = [
  "src/resumex/**",
  "README.md",
  "LICENSE",
]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
